requirements.txt
torch
torchvision
distinctipy
rembg
pillow
numpy
pandas
matplotlib
tqdm
gradio
onnxruntime